Login and registration part implemented

+++ src/main/java/com/example/demo/repository/LoginRepository.java
@@ -1,9 +1,7 @@
 package com.example.demo.repository;
 
+import com.example.demo.entity.Login;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.demo.entity.Login;
-
-public interface LoginRepository extends JpaRepository<Login,Long> {
-
+public interface LoginRepository extends JpaRepository<Login, Integer> {
 }
